requirements.txt
requests
psycopg2-binary
fastapi
uvicorn
asyncpg